backend_tech_to_speak/pyproject.toml
[project]
name = "backend-tech-to-speak"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "google-genai>=1.50.1",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
